JPA oneToMany - user posts, post editor

+++ src/main/java/com/example/MyLittleTwitter/repos/MessageRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Long> {
+public interface MessageRepo extends CrudRepository<Message, Integer> {
 
     List<Message> findByTag(String tag);
